# Websocket handshake perf