--- a/src/Middleware/WebSockets/src/HandshakeHelpers.cs
+++ b/src/Middleware/WebSockets/src/HandshakeHelpers.cs
@@ -66,7 +66,7 @@ namespace Microsoft.AspNetCore.WebSockets
             return validConnection && validUpgrade && validVersion && validKey;
         }
 
-        internal static void GenerateResponseHeaders(string key, string subProtocol, IHeaderDictionary headers)
+        public static void GenerateResponseHeaders(string key, string subProtocol, IHeaderDictionary headers)
         {
             headers[HeaderNames.Connection] = Constants.Headers.ConnectionUpgrade;
             headers[HeaderNames.Upgrade] = Constants.Headers.UpgradeWebSocket;