--- a/src/Middleware/WebSockets/src/HandshakeHelpers.cs
+++ b/src/Middleware/WebSockets/src/HandshakeHelpers.cs
@@ -24,6 +24,15 @@ namespace Microsoft.AspNetCore.WebSockets
             HeaderNames.SecWebSocketVersion
         };
 
+        // "258EAFA5-E914-47DA-95CA-C5AB0DC85B11"
+        private static ReadOnlySpan<byte> _encodedWebSocketKey => new byte[]
+        {
+            (byte)'2', (byte)'5', (byte)'8', (byte)'E', (byte)'A', (byte)'F', (byte)'A', (byte)'5', (byte)'-',
+            (byte)'E', (byte)'9', (byte)'1', (byte)'4', (byte)'-', (byte)'4', (byte)'7', (byte)'D', (byte)'A',
+            (byte)'-', (byte)'9', (byte)'5', (byte)'C', (byte)'A', (byte)'-', (byte)'C', (byte)'5', (byte)'A',
+            (byte)'B', (byte)'0', (byte)'D', (byte)'C', (byte)'8', (byte)'5', (byte)'B', (byte)'1', (byte)'1'
+        };
+
         // Verify Method, Upgrade, Connection, version,  key, etc..
         public static bool CheckSupportedWebSocketRequest(string method, IEnumerable<KeyValuePair<string, string>> headers)
         {
@@ -103,7 +112,7 @@ namespace Microsoft.AspNetCore.WebSockets
         public static string CreateResponseKey(string requestKey)
         {
             // "The value of this header field is constructed by concatenating /key/, defined above in step 4
-            // in Section 4.2.2, with the string "258EAFA5- E914-47DA-95CA-C5AB0DC85B11", taking the SHA-1 hash of
+            // in Section 4.2.2, with the string "258EAFA5-E914-47DA-95CA-C5AB0DC85B11", taking the SHA-1 hash of
             // this concatenated value to obtain a 20-byte value and base64-encoding"
             // https://tools.ietf.org/html/rfc6455#section-4.2.2
 
@@ -114,17 +123,18 @@ namespace Microsoft.AspNetCore.WebSockets
 
             using (var algorithm = SHA1.Create())
             {
-                string merged = requestKey + "258EAFA5-E914-47DA-95CA-C5AB0DC85B11";
-
                 // requestKey is already verified to be small (24 bytes) by 'IsRequestKeyValid()' and everything is 1:1 mapping to UTF8 bytes
                 // so this can be hardcoded to 60 bytes for the requestKey + static websocket string
                 Span<byte> mergedBytes = stackalloc byte[60];
-                Encoding.UTF8.GetBytes(merged, mergedBytes);
+                Encoding.UTF8.GetBytes(requestKey, mergedBytes);
+                _encodedWebSocketKey.CopyTo(mergedBytes.Slice(24));
 
                 Span<byte> hashedBytes = stackalloc byte[20];
                 var success = algorithm.TryComputeHash(mergedBytes, hashedBytes, out var written);
-                Debug.Assert(success);
-                Debug.Assert(written == 20);
+                if (!success || written != 20)
+                {
+                    throw new InvalidOperationException("Could not compute the hash for the 'Sec-WebSocket-Accept' header.");
+                }
 
                 return Convert.ToBase64String(hashedBytes);
             }