--- a/src/Middleware/perf/Microbenchmarks/AssemblyInfo.cs
+++ b/src/Middleware/perf/Microbenchmarks/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[assembly: BenchmarkDotNet.Attributes.AspNetCoreBenchmark]