Found a couple spots we could avoid allocating in the websocket handshake.

Would like @blowdart and @Tratcher to take a look to make sure I'm not doing something completely wrong here.

Before:

Method |        Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
---------------------- |------------:|----------:|----------:|-------------:|-------:|----------:|
CreateResponseKey | 1,118.31 ns | 20.799 ns | 19.455 ns |    894,207.5 | 0.1373 |    5360 B |
IsRequestKeyValid |    92.56 ns |  1.761 ns |  2.163 ns | 10,803,358.7 | 0.0095 |    400 B |

After:

Method |        Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
---------------------- |------------:|----------:|----------:|-------------:|-------:|----------:|
CreateResponseKey |   966.29 ns | 18.022 ns | 15.976 ns |  1,034,883.1 | 0.0763 |    3520 B |
IsRequestKeyValid |    54.61 ns |  1.089 ns |  1.878 ns | 18,311,386.7 |      - |       0 B |

With caching SHA1:

Method |      Mean |    Error |    StdDev |         Op/s |  Gen 0 | Allocated |
------------------ |----------:|---------:|----------:|-------------:|-------:|----------:|
CreateResponseKey | 497.52 ns | 9.885 ns | 10.576 ns |  2,009,951.4 | 0.0153 |     800 B |