@BrennanConroy 2019-07-23
New perf numbers in original comment :D