@BrennanConroy 2019-09-23
🆙 📅 
Removed SHA1 caching to make the PR more enticing :D